fix: compute OpenRouter balance in Decimal space to avoid float precision loss
Decode total_credits/total_usage directly as Decimal instead of Double,
so the remaining-balance subtraction no longer round-trips through
binary floating point before being persisted/displayed.

+++ Sources/OkTally/Plugins/OpenRouter/OpenRouterAPIClient.swift
@@ -3,8 +3,8 @@ import Foundation
 
 struct OpenRouterCreditsResponse: Codable, Equatable {
     struct DataField: Codable, Equatable {
-        let totalCredits: Double
-        let totalUsage: Double
+        let totalCredits: Decimal
+        let totalUsage: Decimal
         enum CodingKeys: String, CodingKey {
             case totalCredits = "total_credits"
             case totalUsage = "total_usage"
